Heights should be floats, not doubles (since NSPoint/NSRect use floats). Added coding/decoding of hidden height. Other experimental stuff (playing with the IB palette).

+++ Frameworks/DisclosableView/SNDisclosableView.h
@@ -4,8 +4,8 @@
 @interface SNDisclosableView : NSView
 {
     BOOL isShown;
-    double originalHeight;
-    double hiddenHeight;
+    float originalHeight;
+    float hiddenHeight;
     NSArray *hiddenSubviews;
     NSView *nonretainedOriginalNextKeyView;
     NSView *nonretainedLastChildKeyView;
@@ -15,8 +15,8 @@
 - (BOOL)isShown;
 - (void)setIsShown:(BOOL)value;
 
-- (double)hiddenHeight;
-- (void)setHiddenHeight:(double)value;
+- (float)hiddenHeight;
+- (void)setHiddenHeight:(float)value;
 
 // Actions
 - (IBAction)toggleDisclosure:(id)sender;
